+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
  - Restore IDTA conformity and reinstate max length of subprotocolBody to 2048
  - High number of SELECT statements during pagination
+ - Remove duplicate `idShort` checks from `ShellService`
 
 ## 0.8.0
 ### Added